Begin converting remaining IFT test data blobs to be_buffer!{}.
