+++ font-test-data/src/lib.rs
@@ -3,6 +3,7 @@
 pub mod gdef;
 pub mod gpos;
 pub mod gsub;
+pub mod ift;
 pub mod layout;
 
 pub static CMAP12_FONT1: &[u8] = include_bytes!("../test_data/ttf/cmap12_font1.ttf");
@@ -218,398 +219,4 @@ pub mod colrv1_json {
             _ => panic!("No entry for {key}, if this is a new test you might need to add a case"),
         }
     }
-}
-
-pub mod ift {
-    use read_fonts::{be_buffer, be_buffer_add, test_helpers::BeBuffer, types::Uint24};
-
-    pub static IFT_BASE: &[u8] = include_bytes!("../test_data/ttf/ift_base.ttf");
-
-    // Format specification: https://w3c.github.io/IFT/Overview.html#patch-map-format-1
-    pub fn simple_format1() -> BeBuffer {
-        let mut buffer = be_buffer! {
-            /* ### Header ### */
-            1u8,                    // format
-            0u32,                   // reserved
-            [1u32, 2, 3, 4],        // compat id
-            2u16,                   // max entry id
-            {2u16: "max_glyph_map_entry_id"},
-            (Uint24::new(7)),       // glyph count
-            {0u32: "glyph_map_offset"},
-            0u32,                   // feature map offset
-            0b00000010u8,           // applied entry bitmap (entry 1)
-
-            8u16,                   // uri template length
-            {b'A': "uri_template[0]"},
-            {b'B': "uri_template[1]"},
-            [b'C', b'D', b'E', b'F', 0xc9, 0xa4], // uri_template[2..7]
-
-            {4u8: "patch_encoding"}, // = glyph keyed
-
-            /* ### Glyph Map ### */
-            {1u16: "glyph_map"},     // first mapped glyph
-            {2u8: "entry_index[1]"},
-            [1u8, 0, 1, 0, 0]        // entry index[2..6]
-        };
-
-        let offset = buffer.offset_for("glyph_map") as u32;
-        buffer.write_at("glyph_map_offset", offset);
-
-        buffer
-    }
-
-    #[rustfmt::skip]
-    pub static U16_ENTRIES_FORMAT1: &[u8] = &[
-        0x01,                    // 0: format
-
-        0x00, 0x00, 0x00, 0x00,  // 1: reserved
-
-        0x00, 0x00, 0x00, 0x01,  // 5: compat id [0]
-        0x00, 0x00, 0x00, 0x02,  // 9: compat id [1]
-        0x00, 0x00, 0x00, 0x03,  // 13: compat id [2]
-        0x00, 0x00, 0x00, 0x04,  // 17: compat id [3]
-
-        0x01, 0x2C,              // 21: max entry id = 300
-        0x01, 0x2C,              // 23: max glyph map entry id = 300
-
-              0x00, 0x00, 0x07,  // 25: glyph count
-        0x00, 0x00, 0x00, 0x55,  // 28: glyph map offset (0x55 = 85)
-        0x00, 0x00, 0x00, 0x00,  // 32: feature map offset
-
-        // 36: applied entry bitmap (38 bytes)
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00,
-
-        0x00, 0x08,              // 74: uriTemplateLength
-        b'A', b'B', b'C', b'D',
-        b'E', b'F', 0xc9, 0xa4,  // 76: uriTemplate[8]
-
-        0x04,                    // 84: patch encoding = glyph keyed
-
-        // ## Glyph Map ##
-        0x00, 0x02,              // 85: first mapped glyph
-        // 88: entryIndex[2..6]
-        0x00, 0x50, // gid 2
-        0x00, 0x51, // gid 3
-        0x01, 0x2c, // gid 4
-        0x01, 0x2c, // gid 5
-        0x00, 0x50, // gid 6
-    ];
-
-    #[rustfmt::skip]
-    pub static FEATURE_MAP_FORMAT1: &[u8] = &[
-        0x01,                    // 0: format
-
-        0x00, 0x00, 0x00, 0x00,  // 1: reserved
-
-        0x00, 0x00, 0x00, 0x01,  // 5: compat id [0]
-        0x00, 0x00, 0x00, 0x02,  // 9: compat id [1]
-        0x00, 0x00, 0x00, 0x03,  // 13: compat id [2]
-        0x00, 0x00, 0x00, 0x04,  // 17: compat id [3]
-
-        0x01, 0x90,              // 21: max entry id = 0x190
-        0x01, 0x2c,              // 23: max glyph map entry id = 0x12C
-        0x00, 0x00, 0x07,        // 25: glyph count
-        0x00, 0x00, 0x00, 0x62,  // 28: glyph map offset (0x62 = 98)
-        0x00, 0x00, 0x00, 0x6E,  // 32: feature map offset (0x6E = 110)
-
-        // 36: applied entry bitmap (51 bytes) - 0x12b is applied
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,       // [0, 64)
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,       // [64, 128)
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,       // [128, 192)
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,       // [192, 256)
-        0x00, 0x00, 0x00, 0x00, 0x00, 0b00001000, 0x00, 0x00, // [256, 320)
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,       // [320, 384)
-        0x00, 0x00, 0x00,                                     // [384, 400)
-
-        0x00, 0x08,              // 87: uriTemplateLength
-        b'A', b'B', b'C', b'D',
-        b'E', b'F', 0xc9, 0xa4,  // 89: uriTemplate[8]
-
-        0x04,                    // 97: patch encoding = glyph keyed
-
-        // ## Glyph Map ##
-        0x00, 0x02,              // 98: first mapped glyph
-        // 101: entryIndex[2..6]
-        0x00, 0x50, // gid 2
-        0x00, 0x51, // gid 3
-        0x01, 0x2c, // gid 4
-        0x01, 0x2b, // gid 5
-        0x00, 0x50, // gid 6
-
-        // ## Feature Map ##
-        0x00, 0x03,               // 110: feature count (3)
-
-        // FeatureRecord[0]
-        b'd', b'l', b'i', b'g',   // 112: feature tag
-        0x01, 0x90,               // 116: first new entry index (0x190 = 400)
-        0x00, 0x01,               // 118: entry map count
-
-        // FeatureRecord[1]
-        b'l', b'i', b'g', b'a',   // 120: feature tag
-        0x01, 0x80,               // 124: first new entry index
-        0x00, 0x02,               // 126: entry map count
-
-        // FeatureRecord[2]
-        b'n', b'u', b'l', b'l',   // 120: feature tag
-        0x01, 0x2D,               // 122: first new entry index
-        0x00, 0x01,               // 124: entry map count
-
-        // EntryMapRecord[0]: "dlig" + entry 0x51 => entry 0x190
-        0x00, 0x51,               // first_entry_index
-        0x00, 0x51,               // last_entry_index
-
-        // EntryMapRecord[1]: "liga" + entry [0x50, 0x51] => entry 0x180
-        0x00, 0x50,               // first_entry_index
-        0x00, 0x51,               // last_entry_index
-
-        // EntryMapRecord[2]: "liga" + entry 0x12c => entry 0x181
-        0x01, 0x2b,               // first_entry_index
-        0x01, 0x2c,               // last_entry_index
-
-        // EntryMapRecord[3]: "null" + entry 0x0 => entry 0x12D
-        0x00, 0x00,               // first_entry_index
-        0x00, 0x00,               // last_entry_index
-    ];
-
-    // Format specification: https://w3c.github.io/IFT/Overview.html#patch-map-format-2
-    #[rustfmt::skip]
-    pub static CODEPOINTS_ONLY_FORMAT2: &[u8] = &[
-        0x02,                    // 0: format
-
-        0x00, 0x00, 0x00, 0x00,  // 1: reserved
-
-        0x00, 0x00, 0x00, 0x01,  // 5: compat id [0]
-        0x00, 0x00, 0x00, 0x02,  // 9: compat id [1]
-        0x00, 0x00, 0x00, 0x03,  // 13: compat id [2]
-        0x00, 0x00, 0x00, 0x04,  // 17: compat id [3]
-
-        0x04,                    // 21: default patch encoding = glyph keyed
-        0x00, 0x00, 0x03,        // 22: entry count
-        0x00, 0x00, 0x00, 0x2b,  // 25: entries offset (0x2b = 43)
-        0x00, 0x00, 0x00, 0x00,  // 29: entry id string data = null
-
-        0x00, 0x08,              // 33: uriTemplateLength
-        b'A', b'B', b'C', b'D',
-        b'E', b'F', 0xc9, 0xa4,  // 35: uriTemplate[8]
-
-        // Entries Array
-        // Entry id = 1
-        0b00010000,                         // 43: format = CODEPOINT_BIT_1
-        0b00001101, 0b00000011, 0b00110001, // 44: codepoints = [0..17]
-
-        // Entry id = 2
-        0b01100000,                         // 47: format = CODEPOINT_BIT_2
-        0x00, 0x05,                         // 48: bias = 6
-        0b00001101, 0b00000011, 0b00110001, // 50: codepoints = [5..22]
-
-        // Entry id = 3
-        0b00100000,                         // 47: format = CODEPOINT_BIT_2
-        0x00, 0x05,                         // 48: bias = 5
-        0b00001101, 0b00000011, 0b00110001, // 50: codepoints = [5..22]
-    ];
-
-    #[rustfmt::skip]
-    pub static FEATURES_AND_DESIGN_SPACE_FORMAT2: &[u8] = &[
-        0x02,                    // 0: format
-
-        0x00, 0x00, 0x00, 0x00,  // 1: reserved
-
-        0x00, 0x00, 0x00, 0x01,  // 5: compat id [0]
-        0x00, 0x00, 0x00, 0x02,  // 9: compat id [1]
-        0x00, 0x00, 0x00, 0x03,  // 13: compat id [2]
-        0x00, 0x00, 0x00, 0x04,  // 17: compat id [3]
-
-        0x04,                    // 21: default patch encoding = glyph keyed
-        0x00, 0x00, 0x03,        // 22: entry count
-        0x00, 0x00, 0x00, 0x2b,  // 25: entries offset (0x2b = 43)
-        0x00, 0x00, 0x00, 0x00,  // 29: entry id string data = null
-
-        0x00, 0x08,              // 33: uriTemplateLength
-        b'A', b'B', b'C', b'D',
-        b'E', b'F', 0xc9, 0xa4,  // 35: uriTemplate[8]
-
-        // Entries Array
-        // Entry id = 1
-        0b00010001,                         // 43: format = CODEPOINT_BIT_1 | FEATURES_AND_DESIGN_SPACE
-
-        0x02,                               // 44: feature count = 2
-        b'l', b'i', b'g', b'a',             // 45: feature[0] = liga
-        b's', b'm', b'c', b'p',             // 49: feature[0] = smcp
-
-        0x00, 0x01,                         // 53: design space count = 1
-        b'w', b'd', b't', b'h',             // 55: tag = wdth
-        0x00, 0x00, 0x80, 0x00,             // 59: start = 0.5
-        0x00, 0x01, 0x00, 0x00,             // 63: end = 1.0
-
-        0b00001101, 0b00000011, 0b00110001, // 67: codepoints = [0..17]
-
-        // Entries Array
-        // Entry id = 2
-        0b00010001,                         // 70: format = CODEPOINT_BIT_1 | FEATURES_AND_DESIGN_SPACE
-
-        0x01,                               // 71: feature count = 1
-        b'r', b'l', b'i', b'g',             // 72: feature[0] = rlig
-
-        0x00, 0x00,                         // 76: design space count = 1
-
-        0b00001101, 0b00000011, 0b00110001, // 78: codepoints = [0..17]
-
-        // Entry id = 3
-        0b000100001,                         // 81: format = CODEPOINT_BIT_2 | FEATURES_AND_DESIGN_SPACE
-
-        0x01,                               // 82: feature count = 1
-        b's', b'm', b'c', b'p',             // 83: feature[0] = smcp
-
-        0x00, 0x03,                         // 87: design space count = 2
-        b'w', b'g', b'h', b't',             // 89: tag = wght
-        0x00, 0xC8, 0x00, 0x00,             // 93: start = 200
-        0x02, 0xBC, 0x00, 0x00,             // 97: end = 700
-
-        b'w', b'd', b't', b'h',             // 101: tag = wdth
-        0x00, 0x00, 0x00, 0x00,             // 105: start = 0
-        0x00, 0x00, 0x80, 0x00,             // 109: end = 0.5
-
-        b'w', b'd', b't', b'h',             // 114: tag = wdth
-        0x00, 0x02, 0x00, 0x00,             // 119: start = 2.0
-        0x00, 0x02, 0x80, 0x00,             // 124: end = 2.5
-
-        0x00, 0x05,                         // 128: bias = 5
-        0b00001101, 0b00000011, 0b00110001, // 130: codepoints = [5..22]
-    ];
-
-    #[rustfmt::skip]
-    pub static COPY_INDICES_FORMAT2: &[u8] = &[
-        0x02,                    // 0: format
-
-        0x00, 0x00, 0x00, 0x00,  // 1: reserved
-
-        0x00, 0x00, 0x00, 0x01,  // 5: compat id [0]
-        0x00, 0x00, 0x00, 0x02,  // 9: compat id [1]
-        0x00, 0x00, 0x00, 0x03,  // 13: compat id [2]
-        0x00, 0x00, 0x00, 0x04,  // 17: compat id [3]
-
-        0x04,                    // 21: default patch encoding = glyph keyed
-        0x00, 0x00, 0x09,        // 22: entry count 9
-        0x00, 0x00, 0x00, 0x2b,  // 25: entries offset (0x2b = 43)
-        0x00, 0x00, 0x00, 0x00,  // 29: entry id string data = null
-
-        0x00, 0x08,              // 33: uriTemplateLength
-        b'A', b'B', b'C', b'D',
-        b'E', b'F', 0xc9, 0xa4,  // 35: uriTemplate[8]
-
-        // Entries Array
-
-        // Entry id = 1
-        0b00100000,                         // : format = CODEPOINT_BIT_2
-        0x00, 0x05,                         // : bias = 5
-        0b00001101, 0b00000011, 0b00110001, // : codepoints = [5..22]
-
-        // Entry id = 2
-        0b00100000,                         // : format = CODEPOINT_BIT_2
-        0x00, 0x32,                         // : bias = 50
-        0b00001101, 0b00000011, 0b00110001, // : codepoints = [50..67]
-
-        // Entry id = 3
-        0b00000001,                         //   : format = FEATURES_AND_DESIGN_SPACE
-
-        0x01,                               //   : feature count = 1
-        b'r', b'l', b'i', b'g',             //   : feature[0] = rlig
-
-        0x00, 0x01,                         //   : design space count = 1
-        b'w', b'g', b'h', b't',             //   : tag = wght
-        0x00, 0xC8, 0x00, 0x00,             //   : start = 200
-        0x02, 0xBC, 0x00, 0x00,             //   : end = 700
-
-        // Entry id = 4
-        0b00000001,                         //   : format = FEATURES_AND_DESIGN_SPACE
-
-        0x01,                               //   : feature count = 1
-        b'l', b'i', b'g', b'a',             //   : feature[0] = liga
-
-        0x00, 0x01,                         //   : design space count = 1
-        b'w', b'g', b'h', b't',             //   : tag = wght
-        0x00, 0x32, 0x00, 0x00,             //   : start = 50
-        0x00, 0x64, 0x00, 0x00,             //   : end = 100
-
-        // Entry id = 5
-        0b00000010,                         //   : format = COPY_INDICES
-        0x01,                               //   : copy count = 1
-        0x00, 0x00, 0x00,                   //   : copy 0
-
-        // Entry id = 6
-        0b00000010,                         //   : format = COPY_INDICES
-        0x01,                               //   : copy count = 1
-        0x00, 0x00, 0x02,                   //   : copy 2
-
-        // Entry id = 7
-        0b00000010,                         //   : format = COPY_INDICES
-        0x04,                               //   : copy count = 4
-        0x00, 0x00, 0x03,                   //   : copy 3
-        0x00, 0x00, 0x02,                   //   : copy 2
-        0x00, 0x00, 0x01,                   //   : copy 1
-        0x00, 0x00, 0x00,                   //   : copy 0
-
-        // Entry id = 8
-        0b00000010,                         //   : format = COPY_INDICES
-        0x02,                               //   : copy count = 2
-        0x00, 0x00, 0x04,                   //   : copy 4
-        0x00, 0x00, 0x05,                   //   : copy 5
-
-        // Entry id = 9
-        0b00100010,                         // : format = CODEPOINT_BIT_2 | COPY_INDICES
-        0x01,                               // : copy count = 1
-        0x00, 0x00, 0x00,                   // : copy 0
-        0x00, 0x64,                         // : bias = 100
-        0b00001101, 0b00000011, 0b00110001, // : codepoints = [100..117]
-    ];
-
-    // Format specification: https://w3c.github.io/IFT/Overview.html#patch-map-format-2
-    #[rustfmt::skip]
-    pub static CUSTOM_IDS_FORMAT_2: &[u8] = &[
-        0x02,                    // 0: format
-
-        0x00, 0x00, 0x00, 0x00,  // 1: reserved
-
-        0x00, 0x00, 0x00, 0x01,  // 5: compat id [0]
-        0x00, 0x00, 0x00, 0x02,  // 9: compat id [1]
-        0x00, 0x00, 0x00, 0x03,  // 13: compat id [2]
-        0x00, 0x00, 0x00, 0x04,  // 17: compat id [3]
-
-        0x04,                    // 21: default patch encoding = glyph keyed
-        0x00, 0x00, 0x04,        // 22: entry count
-        0x00, 0x00, 0x00, 0x2b,  // 25: entries offset (0x2b = 43)
-        0x00, 0x00, 0x00, 0x00,  // 29: entry id string data = null
-
-        0x00, 0x08,              // 33: uriTemplateLength
-        b'A', b'B', b'C', b'D',
-        b'E', b'F', 0xc9, 0xa4,  // 35: uriTemplate[8]
-
-        // Entries Array
-        // Entry id = 0
-        0b00010100,                         // 43: format = CODEPOINT_BIT_1 | ID_DELTA
-        0xFF, 0xFF, 0xFF,                   // 44: id delta = -1
-        0b00001101, 0b00000011, 0b00110001, // 47: codepoints = [0..17]
-
-        // Entry id = 6
-        0b00100100,                         // 50: format = CODEPOINT_BIT_2 | ID_DELTA
-        0x00, 0x00, 0x05,                   // 51: id delta 5
-        0x00, 0x05,                         // 54: bias = 6
-        0b00001101, 0b00000011, 0b00110001, // 56: codepoints = [5..22]
-
-        // Entry id = 14
-        0b01000100,                         // 59: format = ID_DELTA | IGNORED
-        0x00, 0x00, 0x07,                   // 60: id delta 7
-
-        // Entry id = 15
-        0b00101000,                         // 63: format = CODEPOINT_BIT_2 | PATCH_ENCODING
-        0x04,                               // 64: patch encoding = Glyph Keyed
-        0x00, 0x0A,                         // 65: bias = 10
-        0b00001101, 0b00000011, 0b00110001, // 67: codepoints = [10..27]
-    ];
 }
